requirements.txt
boto3==1.4.4
Jinja2==2.9.6
figure-sdk==0.2.4
pycups==1.9.73
pypdf2==1.26.0

